Export rb_yjit_tracing_invalidate_all() from Rust
Swaps btest crashes with a different one.
