+++ yjit/src/codegen.rs
@@ -5100,111 +5100,6 @@ fn gen_opt_invokebuiltin_delegate(jit: &mut JITState, ctx: &mut Context, cb: &mu
 
     KeepCompiling
 }
-
-// Invalidate all generated code and patch C method return code to contain
-// logic for firing the c_return TracePoint event. Once rb_vm_barrier()
-// returns, all other ractors are pausing inside RB_VM_LOCK_ENTER(), which
-// means they are inside a C routine. If there are any generated code on-stack,
-// they are waiting for a return from a C routine. For every routine call, we
-// patch in an exit after the body of the containing VM instruction. This makes
-// it so all the invalidated code exit as soon as execution logically reaches
-// the next VM instruction. The interpreter takes care of firing the tracing
-// event if it so happens that the next VM instruction has one attached.
-//
-// The c_return event needs special handling as our codegen never outputs code
-// that contains tracing logic. If we let the normal output code run until the
-// start of the next VM instruction by relying on the patching scheme above, we
-// would fail to fire the c_return event. The interpreter doesn't fire the
-// event at an instruction boundary, so simply exiting to the interpreter isn't
-// enough. To handle it, we patch in the full logic at the return address. See
-// full_cfunc_return().
-//
-// In addition to patching, we prevent future entries into invalidated code by
-// removing all live blocks from their iseq.
-void
-rb_yjit_tracing_invalidate_all(void)
-{
-    if (!rb_yjit_enabled_p()) return;
-
-    // Stop other ractors since we are going to patch machine code.
-    RB_VM_LOCK_ENTER();
-    rb_vm_barrier();
-
-    // Make it so all live block versions are no longer valid branch targets
-    rb_objspace_each_objects(tracing_invalidate_all_i, NULL);
-
-    // Apply patches
-    const uint32_t old_pos = cb->write_pos;
-    rb_darray_for(global_inval_patches, patch_idx) {
-        struct codepage_patch patch = rb_darray_get(global_inval_patches, patch_idx);
-        cb.set_pos(patch.inline_patch_pos);
-        uint8_t *jump_target = cb_get_ptr(ocb, patch.outlined_target_pos);
-        jmp_ptr(cb, jump_target);
-    }
-    cb.set_pos(old_pos);
-
-    // Freeze invalidated part of the codepage. We only want to wait for
-    // running instances of the code to exit from now on, so we shouldn't
-    // change the code. There could be other ractors sleeping in
-    // branch_stub_hit(), for example. We could harden this by changing memory
-    // protection on the frozen range.
-    RUBY_ASSERT_ALWAYS(yjit_codepage_frozen_bytes <= old_pos && "frozen bytes should increase monotonically");
-    yjit_codepage_frozen_bytes = old_pos;
-
-    cb_mark_all_executable(ocb);
-    cb_mark_all_executable(cb);
-    RB_VM_LOCK_LEAVE();
-}
-
-static int
-tracing_invalidate_all_i(void *vstart, void *vend, size_t stride, void *data)
-{
-    VALUE v = (VALUE)vstart;
-    for (; v != (VALUE)vend; v += stride) {
-        void *ptr = asan_poisoned_object_p(v);
-        asan_unpoison_object(v, false);
-
-        if (rb_obj_is_iseq(v)) {
-            rb_iseq_t *iseq = (rb_iseq_t *)v;
-            invalidate_all_blocks_for_tracing(iseq);
-        }
-
-        asan_poison_object_if(ptr, v);
-    }
-    return 0;
-}
-
-static void
-invalidate_all_blocks_for_tracing(const rb_iseq_t *iseq)
-{
-    struct rb_iseq_constant_body *body = iseq->body;
-    if (!body) return; // iseq yet to be initialized
-
-    ASSERT_vm_locking();
-
-    // Empty all blocks on the iseq so we don't compile new blocks that jump to the
-    // invalidted region.
-    // TODO Leaking the blocks for now since we might have situations where
-    // a different ractor is waiting in branch_stub_hit(). If we free the block
-    // that ractor can wake up with a dangling block.
-    rb_darray_for(body->yjit_blocks, version_array_idx) {
-        rb_yjit_block_array_t version_array = rb_darray_get(body->yjit_blocks, version_array_idx);
-        rb_darray_for(version_array, version_idx) {
-            // Stop listening for invalidation events like basic operation redefinition.
-            block_t *block = rb_darray_get(version_array, version_idx);
-            yjit_unlink_method_lookup_dependency(block);
-            yjit_block_assumptions_free(block);
-        }
-        rb_darray_free(version_array);
-    }
-    rb_darray_free(body->yjit_blocks);
-    body->yjit_blocks = NULL;
-
-#if USE_MJIT
-    // Reset output code entry point
-    body->jit_func = NULL;
-#endif
-}
 */
